requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.55.3
numpy==1.24.3
pydantic==2.5.1
python-dotenv==1.0.0
requests==2.32.2